+++ test/unit-test.cjs
@@ -1279,7 +1279,7 @@ const test = () => {
   new Expect(img, trustSettings()).toNotBuild();
   new Expect(img).toParse(trustSettings()) // no trust
   markup = temml.renderToString(img, trustSettings()).replace(mathTagRegEx, "");
-  new Expect(markup).toBe(`<mtext><img src='https://cdn.kastatic.org/images/apple-touch-icon-57x57-precomposed.new.png' alt='KA logo' style="height:0.9em;width:0.9em;"/></mtext>`);
+  new Expect(markup).toBe(`<mtext><img src='https://cdn.kastatic.org/images/apple-touch-icon-57x57-precomposed.new.png' alt='KA logo' style="height:0.9em;width:0.9em;"></mtext>`);
 
   assertion = "An HTML extension builder should work"
   const html = r`\id{bar}{x}\class{foo}{x}\style{color: red;}{x}\data{foo=a, bar=b}{x}`;